Crate nodrop_union

nodrop-union is the untagged unions (requires Rust nightly channel and unstable as of this writing) implementation for the nodrop crate.

It is intended you use this through the nodrop crate with the use_union crate feature enabled.

This is the future implementation of nodrop, once it is stable.

This implementation is a lot better:

Does not have a destructor at all

Can be Copy if T is Copy

No space overhead / no runtime flag

This means that this implementation has extensions that the stable nodrop does not yet have, which is something to be aware of if you are switching.

Structs

NoDrop
A type holding T that will not call its destructor on drop