Struct nodrop_union::NoDrop

pub struct NoDrop<T>(_);

A type holding T that will not call its destructor on drop

The untagged unions implementation of NoDrop<T> is Copy where T: Copy, which was not possible in the stable implementation.

Methods

impl<T> NoDrop<T>

fn new(value: T) -> Self

Create a new NoDrop.

fn into_inner(self) -> T

Extract the inner value.

Once extracted, the value can of course drop again.